+++ src/Endpoint/BulkSignatureEndpoint.php
@@ -6,6 +6,7 @@ namespace DigitalCz\DigiSign\Endpoint;
 
 use DigitalCz\DigiSign\DigiSign;
 use DigitalCz\DigiSign\Endpoint\Traits\GetEndpointTrait;
+use DigitalCz\DigiSign\Endpoint\Traits\ListEndpointTrait;
 use DigitalCz\DigiSign\Resource\BulkSignature;
 
 /**
@@ -14,6 +15,8 @@ use DigitalCz\DigiSign\Resource\BulkSignature;
 final class BulkSignatureEndpoint extends ResourceEndpoint
 {
     use GetEndpointTrait;
+    /** @use ListEndpointTrait<BulkSignature> */
+    use ListEndpointTrait;
 
     public function __construct(DigiSign $parent)
     {